CPU: wait before running the instruction
I noticed that with some PPU tests, the VBlank flag for example will be
set in the middle of the instruction execution.
For example if an instruction takes 4 CPU clocks, that means 12 PPU
clock, and thats a lot of time for some flags to change over.

This commit tries to solve this by instead of executing the instruction
and then waiting for the appropriate cycles, first wait and then
execute, in this case since some instructions will wait more depending
on the execution itself (like branch instructions), just wait the base
time, which is the minimum time for that instruction.

Next is to try to fix the PPU tests