NES + tester: fix, CPU clock should be run before PPU
There is not really a reason for this, after all in the hardware, the
two clocks will run in parallel, but this way it actually fix some
tests, and seems more stable.

Also for the `nes_ui_base/src/nes.rs`, before it was running 2 CPU
clocks and then 6 PPU clocks, which is so NO NO NO (the tests were
acting so weird). So fixed it