CPU: code improvement for `instruction.rs`
The most major improvement is removing the array in
`get_instruction_len`, instead, a reference array can be used like
`&[..]`, or a match block like it was used. The reason for this
improvement is that using plain array `[..]`, will generate the array
inside the stack of the function, and it will be generated every time
that function is called. Using `&[..]` or match block will be optimized
and stored as an array in `.rodata` binary section.

Also removed the return type `Result` from Instruction::from_byte, as
now we can generate all instructions (including unofficial), this will
not return `Err`. Maybe thinking of putting a flag or something to
indicate that the instruction is illegal(unofficial).