Cartridge: major rewrite to how mapping is done
The best method for this is to have an array of references/pointers to
the banks which should be used, they could have been in place of
`BankMapping` in for example `cpu_ram_memory_mapping`. Very easy to
access, but with rust normal references cannot be used, as you would
have to keep the reference for the entire life of the mapper, and you
need it in mutable and immutable states. So it will not work.