APU: enhanced APU clock accuracy, and added tests
Now APU's `clock` function should be called with the same speed as `CPU`
and it will handle itself, the reason for this is because some things in
the APU runs in odd clocks or even clocks, and some stuff run in every
CPU clock, like the triangle timer.

Together with this, also added some tests, most pass but there are still
some problems with the timing of length counters halt and reloads.

Also Mario's problem is not fixed, even after some tests passing :(,
mostly it has something to do with the upper mentioned not passing tests