NES: fixed windows slow emulation (was due to windows Sleep function)
`std :: thread :: sleep` is not the same in all platforms.
Rust uses `libc::nanosleep` in Linux which can take a number of nano seconds,
sleeping may not be exactly the time it was requested but at least
close.

On the other hand, Rust uses windows api function `Sleep`, it takes a
number of **Milliseconds**, this is not very good for small sleeps,
but this is what I want, and what the emulator was doing the whole time.

Changed the number of clocks to a whole frame, now that the APU is not
very affected by the sleeping time like before, this is not much of a
problem and now the emulator can run smoothly in windows. :)

*they should fix their Kernal*.