NES: fix reset/load_rom cause delay in audio
This is caused by the great amount of cycles spent without any update
about `apu_freq`, so the delay is great. I think the delay was also
there before, but because it was a small number of cycles it was not
noticed.

This is fixed by adding
```
self.apu.borrow_mut().update_apu_freq(1.7 * 1E6 / 2.)
```
After reseting.

This commit adds this into a macro to be able to copy it into multiple
places, maybe using function is better, but rust's design prefer
copying, so a little more won't hurt.

Also it might be optimized into something else.