Mobile + TV: bugfix, android x86 crash due to huge stack size from TV
The TV stored the buffer in the stack, which is 256 * 240 * 3 bytes == 184320
Which results in SEGV from __rust_probestack. Now its in the heap, so no
problems

Also a bug where dart ffigen would convert `run_nes(i64)` into Int32,
becuase `int64_t` is not being imported without "stdint.h", so its fixed
now. This problem will not pass the correct value of `send_port`, as the
port is 64bit integer, on 64bit systems, its ok, but in 32bit
`eax` and `edx` are used together to pass 64bit integer value.