Mobile: Audio into mobile
Used `flutter_sound` library which require at least android API 18.

The library can only accept i16 sampling, so we convert f32 into i16.

The performance is not the best, as when we have control over the apu,
we can detect if the buffer is overusing or underusing and edit
sample recording accordingly. We need to fix this.