Mobile + NES: better frame limiter to fix audio delay
When playing audio locally (using rodio), we can adjust how many
samples we can take in the apu itself to get an audio stream without
delays or problems, which is the best option.
But when sending the audio buffer to an outside audio player (like the
mobile application), we cannot know if we are underusing or overusing
our audio buffer.

The solution I used is to ensure the frame limit stayed fixed at all
time as much as possible